+++ src/main/java/carrotbat410/lol/repository/UserRepository.java
@@ -3,10 +3,9 @@ package carrotbat410.lol.repository;
 import carrotbat410.lol.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Boolean existsByUsername(String username);
 
-    //TODO Entity반환X
     User findByUsername(String username);
 }
